gui/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quadcast-rgb-gui"
version = "0.1.0"
description = "Qt6 GUI for HyperX QuadCast RGB lighting control"
license = {text = "GPL-2.0-only"}
requires-python = ">=3.11"
dependencies = ["PySide6>=6.6", "libusb1>=3.0"]

[project.scripts]
quadcast-rgb-gui = "quadcast_rgb.main:main"

[tool.setuptools.packages.find]
include = ["quadcast_rgb*"]
